Replace Signed Data Type Instead of Unsigned in File Hierarchy

+++ FileEnv/AudioParameters.cs
@@ -10,18 +10,18 @@ namespace FileEnv
 {
     public abstract class AudioParameters : PlayTime, IAudioParameters
     {
-        private ulong _bitRate; // bps
-        private ulong _sampleRate; // Hz
+        private int _bitRate; // bps
+        private int _sampleRate; // Hz
 
-        public AudioParameters(string type, string name, ulong size, string location,
-            string length, ulong bitRate, ulong sampleRate) :
+        public AudioParameters(string type, string name, int size, string location,
+            string length, int bitRate, int sampleRate) :
             base(type, name, size, location, length)
         {
             BitRate = bitRate;
             SampleRate = sampleRate;
         }
 
-        public ulong BitRate { get => _bitRate; set => _bitRate = value; }
-        public ulong SampleRate { get => _sampleRate; set => _sampleRate = value; }
+        public int BitRate { get => _bitRate; set => _bitRate = value; }
+        public int SampleRate { get => _sampleRate; set => _sampleRate = value; }
     }
 }
