+++ FileEnv/Photo.cs
@@ -12,11 +12,11 @@ namespace FileEnv
     public class Photo : File, IPhoto
     {
         private string _dimensions;
-        private ushort _width;
-        private ushort _height;
+        private int _width;
+        private int _height;
 
-        public Photo(string type, string name, ulong size, string location, // File
-            string dimensions, ushort width, ushort height) : // Photo
+        public Photo(string type, string name, int size, string location, // File
+            string dimensions, int width, int height) : // Photo
             base(type, name, size, location)
         {
             Dimensions = dimensions;
@@ -25,8 +25,8 @@ namespace FileEnv
         }
 
         public string Dimensions { get => _dimensions; set => _dimensions = value; }
-        public ushort Width { get => _width; set => _width = value; }
-        public ushort Height { get => _height; set => _height = value; }
+        public int Width { get => _width; set => _width = value; }
+        public int Height { get => _height; set => _height = value; }
 
         public override void Accept(IVisitor visitor)
         {
