+++ FileEnv/Video.cs
@@ -11,14 +11,14 @@ namespace FileEnv
 {
     public class Video : AudioParameters, IVideo
     {
-        private ushort _frameWidth;
-        private ushort _frameHeight;
-        private ushort _frameRate;
+        private int _frameWidth;
+        private int _frameHeight;
+        private int _frameRate;
 
-        public Video(string type, string name, ulong size, string location, // File
-            ulong bitRate, ulong sampleRate, // AudioParameters
+        public Video(string type, string name, int size, string location, // File
+            int bitRate, int sampleRate, // AudioParameters
             string length, // PlayTime
-            ushort frameWidth, ushort frameHeight, ushort frameRate) : // Video
+            int frameWidth, int frameHeight, int frameRate) : // Video
             base(type, name, size, location, length, bitRate, sampleRate)
         {
             FrameWidth = frameWidth;
@@ -26,9 +26,9 @@ namespace FileEnv
             FrameRate = frameRate;
         }
 
-        public ushort FrameWidth { get => _frameWidth; set => _frameWidth = value; }
-        public ushort FrameHeight { get => _frameHeight; set => _frameHeight = value; }
-        public ushort FrameRate { get => _frameRate; set => _frameRate = value; }
+        public int FrameWidth { get => _frameWidth; set => _frameWidth = value; }
+        public int FrameHeight { get => _frameHeight; set => _frameHeight = value; }
+        public int FrameRate { get => _frameRate; set => _frameRate = value; }
 
         public override void Accept(IVisitor visitor)
         {
